# Conflicts:
#	shotgun-client/app/App.js
#	shotgun-client/app/common/components/checkout/DeliveryMap.js
#	shotgun-client/app/common/dao/OrderSummaryDao.js
#	shotgun-client/app/customer/CustomerOrderDetail.js
#	shotgun-client/app/customer/CustomerOrderInProgress.js

+++ viewserver-shotgun/src/main/java/com/shotgun/viewserver/order/Product.java
@@ -5,13 +5,13 @@ package com.shotgun.viewserver.order;
  */
 public class Product {
     String productId,name,description,categoryId;
-    Double price;
+    int price;
 
-    public Double getPrice() {
+    public int getPrice() {
         return price;
     }
 
-    public void setPrice(Double price) {
+    public void setPrice(int price) {
         this.price = price;
     }
