+++ viewserver-shotgun/src/main/java/com/shotgun/viewserver/setup/datasource/ProductDataSource.java
@@ -23,7 +23,7 @@ ProductDataSource {
                         new Column("name", "name", ColumnType.String),
                         new Column("description", "description", ColumnType.String),
                         new Column("categoryId", "categoryId", ColumnType.String),
-                        new Column("price", "price", ColumnType.Double),
+                        new Column("price", "price", ColumnType.Int),
                         new Column("imageUrl", "imageUrl", ColumnType.String)
                 ))
                 .withKeyColumns("productId");
